Deployer processor: Stale resource cleanup Deploy volume processing Undeploy volume processing
