+++ commons/src/main/java/io/hyscale/commons/listener/HyscaleStartUpListener.java
@@ -20,14 +20,12 @@ import org.springframework.context.ApplicationContext;
 import org.springframework.context.ApplicationContextAware;
 import org.springframework.stereotype.Component;
 
-import io.hyscale.commons.utils.HyscaleContextUtil;
-
 @Component
 public class HyscaleStartUpListener implements ApplicationContextAware {
 
     @Override
     public void setApplicationContext(ApplicationContext applicationContext) throws BeansException {
-        HyscaleContextUtil.setContext(applicationContext);
+        HyscaleContextHelper.setContext(applicationContext);
     }
 
 }
