+++ deployer/deployer-services/src/main/java/io/hyscale/deployer/services/aspect/DeployerAspect.java
@@ -28,8 +28,8 @@ import org.springframework.context.annotation.Configuration;
 import io.hyscale.commons.annotations.ComponentInterceptor;
 import io.hyscale.commons.component.IInterceptorProcessor;
 import io.hyscale.commons.exception.HyscaleException;
+import io.hyscale.commons.listener.HyscaleContextHelper;
 import io.hyscale.commons.models.DeploymentContext;
-import io.hyscale.commons.utils.HyscaleContextUtil;
 import io.hyscale.deployer.services.processor.DeployerInterceptorProcessor;
 
 @Aspect
@@ -82,14 +82,14 @@ public class DeployerAspect {
         }
     }
 
-    @AfterThrowing("deploymentContextPointCut(interceptor, context)")
-    public void onError(JoinPoint jp, ComponentInterceptor interceptor, DeploymentContext context)
+    @AfterThrowing(pointcut = "deploymentContextPointCut(interceptor, context)", throwing = "th")
+    public void onError(JoinPoint jp, ComponentInterceptor interceptor, DeploymentContext context, Throwable th)
             throws HyscaleException {
         try {
             for (Class<? extends IInterceptorProcessor> processor : interceptor.processors()) {
                 DeployerInterceptorProcessor deployerProcessor = validateAndGetProcessorBean(processor);
                 if (deployerProcessor != null) {
-                    deployerProcessor.onError(context);
+                    deployerProcessor.onError(context, th);
                 }
             }
         } catch (HyscaleException ex) {
@@ -99,7 +99,7 @@ public class DeployerAspect {
     }
 
     private DeployerInterceptorProcessor validateAndGetProcessorBean(Class<? extends IInterceptorProcessor> processor) {
-        IInterceptorProcessor processorBean = HyscaleContextUtil.getSpringBean(processor);
+        IInterceptorProcessor processorBean = HyscaleContextHelper.getSpringBean(processor);
         if (processorBean == null) {
             logger.debug("Bean not found for Processor {}, ignoring processing", processor.getCanonicalName());
             return null;
